Unificado dos os arquivos de listener em logs, e adicionado nova função de logs
