--- src/main/java/starcode/ListenerAdapterAdapter.java
@@ -1,5 +1,0 @@
-package starcode;
-
-public class ListenerAdapterAdapter {
-
-}
