+++ src/main/java/starcode/PingCommand.java
